requirements.txt
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.12.2
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
numpy==1.24.3
omegaconf==2.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pytest==7.3.1
pyrootutils==1.0.4
pytorch-lightning==2.0.2
PyYAML==6.0
rich==13.3.5
scipy==1.10.1
six==1.16.0
stribor==0.1.0
sympy==1.12
timm==0.9.5
tomli==2.0.1
torch==2.0.1
torchdiffeq==0.2.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
wandb==0.15.8